terms and conditions to separate activity
