+++ app/src/main/java/nl/nfi/cellscanner/ViewMeasurementsActivity.java
@@ -14,6 +14,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.text.Html;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
@@ -69,8 +70,14 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (!hasUserConsent(this)) showTermsAndConditionsScreen();
-        else showRecorderScreen();
+        showRecorderScreen();
+        AppInfoActivity.showIfNoConsent(this);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        AppInfoActivity.showIfNoConsent(this);
     }
 
     @Override
@@ -85,67 +92,6 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
         SharedPreferences sharedPreferences = this.getSharedPreferences(CellScannerApp.TITLE, MODE_PRIVATE);
         sharedPreferences.unregisterOnSharedPreferenceChangeListener(this);
         super.onDestroy();
-    }
-
-    private void showTermsAndConditionsScreen() {
-        setContentView(nl.nfi.cellscanner.R.layout.terms_and_conditions);
-
-        final Context context = this;
-
-        // Show the android ID to the user
-        String androidId = Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ANDROID_ID);
-        TextView tvAndroidId = findViewById(R.id.tac_android_id_text);
-        tvAndroidId.setText(androidId);
-
-        // Set textview with string containing HTML for URLs and phonenumbers
-        String contactInformation = getString(R.string.tac_contact_information);
-        String dataCollection = getString(R.string.tac_data_collection);
-        String aboutApp = getString(R.string.tac_about_app);
-
-        TextView tvContactInfo = findViewById(R.id.tac_contactInformation_text);
-        TextView tvDataCollection = findViewById(R.id.tac_dataCollected_text);
-        TextView tvAboutApp = findViewById(R.id.tac_aboutApp_text);
-
-        setHtmlVersionDependent(contactInformation, tvContactInfo);
-        setHtmlVersionDependent(dataCollection, tvDataCollection);
-        setHtmlVersionDependent(aboutApp, tvAboutApp);
-
-        boolean userConsent = hasUserConsent(context);
-
-        final Button close_button = findViewById(R.id.tac_close_button);
-        close_button.setEnabled(userConsent);
-
-        /*
-        Only allow (un)checking the consent the first time around. To retract consent, an email
-        should be send asking to remove all data.
-         */
-        final CheckBox accepted_checkbox = findViewById(R.id.tac_checkbox);
-        accepted_checkbox.setChecked(userConsent);
-        if (!userConsent) {
-            accepted_checkbox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                @Override
-                public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                    setUserConsent(context, isChecked);
-                    close_button.setEnabled(isChecked);  // Enabled if agreed to T & C
-                }
-            });
-        } else {
-            accepted_checkbox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                @Override
-                public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                    accepted_checkbox.setChecked(true);
-                    Toast.makeText(context, "To revoke consent, contact the NFI to remove your data and then uninstall the app", Toast.LENGTH_SHORT).show();
-                }
-            });
-        }
-    }
-
-    private void setHtmlVersionDependent(String htmlString, TextView tvObject) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            tvObject.setText(Html.fromHtml(htmlString, Html.FROM_HTML_MODE_LEGACY));
-        } else {
-            tvObject.setText(Html.fromHtml(htmlString));
-        }
     }
 
     private void showRecorderScreen() {
@@ -234,19 +180,10 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
 
 
     /**
-     * When closing terms and conditions screen, load the recorder screen
-     *
-     * @param view
-     */
-    public void closeTermsAndConditions(View view) {
-        showRecorderScreen();
-    }
-
-    /**
      * Open the terms and conditions screen on clicking the Info button
      */
     public void openTermsAndConditions(View view) {
-        showTermsAndConditionsScreen();
+        AppInfoActivity.show(this);
     }
 
     /**
